working parser module
