+++ mandatory/lexer/lexer.c
@@ -6,7 +6,7 @@
 /*   By: mamesser <mamesser@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/06 16:15:37 by mamesser          #+#    #+#             */
-/*   Updated: 2023/08/09 15:12:37 by mamesser         ###   ########.fr       */
+/*   Updated: 2023/08/10 15:49:05 by mamesser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
