+++ mandatory/utils/linked_lst.c
@@ -6,7 +6,7 @@
 /*   By: mamesser <mamesser@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/06 16:03:16 by mamesser          #+#    #+#             */
-/*   Updated: 2023/08/10 13:21:44 by mamesser         ###   ########.fr       */
+/*   Updated: 2023/08/10 16:16:08 by mamesser         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,9 +47,7 @@ void	free_cmd_and_lex(t_cmd **cmd_lst, t_lex **lex_lst)
 void	ft_cmd_lst_free(t_cmd **lst)
 {
 	t_cmd	*temp;
-	int		i;
 
-	i = -1;
 	while (*lst)
 	{
 		temp = *lst;
@@ -58,12 +56,8 @@ void	ft_cmd_lst_free(t_cmd **lst)
 			free(temp->cmd);
 		if (temp->path)
 			free(temp->path);
-		// if (temp->args)
-		// {
-		// 	while (temp->args[++i])
-		// 		free(temp->args[i]);
-		// 	free(temp->args);
-		// }
+		if (temp->args)
+			ft_free(temp->args);
 		if (temp->delim)
 			free(temp->delim);
 		if (temp->file)
